Make `in_data.current_time*` f64

+++ after-effects/src/pf/in_data.rs
@@ -106,14 +106,14 @@ impl InData {
     pub fn height(&self) -> i32 {
         unsafe { (*self.ptr).height }
     }
-    pub fn current_frame(&self) -> f32 {
-        unsafe { (*self.ptr).current_time as f32 / (*self.ptr).time_step as f32 }
+    pub fn current_frame(&self) -> f64 {
+        unsafe { (*self.ptr).current_time as f64 / (*self.ptr).time_step as f64 }
     }
-    pub fn current_frame_local(&self) -> f32 {
-        unsafe { (*self.ptr).current_time as f32 / (*self.ptr).local_time_step as f32 }
+    pub fn current_frame_local(&self) -> f64 {
+        unsafe { (*self.ptr).current_time as f64 / (*self.ptr).local_time_step as f64 }
     }
-    pub fn current_timestamp(&self) -> f32 {
-        unsafe { (*self.ptr).current_time as f32 / (*self.ptr).time_scale as f32 }
+    pub fn current_timestamp(&self) -> f64 {
+        unsafe { (*self.ptr).current_time as f64 / (*self.ptr).time_scale as f64 }
     }
     pub fn current_time(&self) -> i32 {
         unsafe { (*self.ptr).current_time }
